+++ sdk/network/Azure.ResourceManager.Network/tests/Helpers/NetworkServiceClientTestBase.cs
@@ -667,10 +667,7 @@ namespace Azure.ResourceManager.Network.Tests.Helpers
                 {
                     AddressPrefixes = { "10.0.0.0/16", }
                 },
-                DhcpOptions = new DhcpOptions()
-                {
-                    DnsServers = { "10.1.1.1", "10.1.2.4" }
-                },
+                DhcpOptionsDnsServers = { "10.1.1.1", "10.1.2.4" },
                 Subnets = { new SubnetData() { AddressPrefix = "10.0.0.0/24", } }
             };
 
@@ -691,10 +688,7 @@ namespace Azure.ResourceManager.Network.Tests.Helpers
                 {
                     AddressPrefixes = { "10.0.0.0/16", }
                 },
-                DhcpOptions = new DhcpOptions()
-                {
-                    DnsServers = { "10.1.1.1", "10.1.2.4" }
-                },
+                DhcpOptionsDnsServers = { "10.1.1.1", "10.1.2.4" },
                 Subnets = { new SubnetData() { AddressPrefix = "10.0.0.0/24", } }
             };